fix: roll ricocheting wall-nuts across lanes smoothly instead of snapping

+++ core/src/main/java/views/battle/Lawn.java
@@ -77,7 +77,7 @@ public final class Lawn {
         return bottom + (float) (GameSession.ROWS - row) * cellHeight;
     }
 
-    public static float rowCenter(int row) {
+    public static float rowCenter(double row) {
         return rowBottom(row) + cellHeight / 2f;
     }
